Issue #172 refactor: Mapper 패키지
